refactor(dashboard): extract SubscriptionVerifyGroup into its own file
